controller/requirements.txt
streamdeck
Pillow
requests
Flask
wakeonlan
websocket-client
websockets
